% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseToGenes.R
\name{collapseToGenes}
\alias{collapseToGenes}
\title{Collapse transcript_ids or gene_ids to gene name level}
\usage{
collapseToGenes(tpm, geneName.map = LGPclassifiers::geneName.map)
}
\arguments{
\item{tpm}{matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{geneName.map}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}
}
\description{
Collapse transcript_ids or gene_ids to gene name level
}
