% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCOO.R
\name{computeCOO}
\alias{computeCOO}
\title{Classify Sample Cell Of Origin}
\usage{
computeCOO(
  query,
  useReference = TRUE,
  reference = NULL,
  id2geneName = NULL,
  featureType = "gene_id",
  ...
)
}
\arguments{
\item{query}{Matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{useReference}{Whether to use an external reference to perform a single sample classification. (default T)}

\item{reference}{Optional matrix of reference input samples (values must be in TPM, one row per
transcript/gene id). If not specified, will default to Robust reference dataset.}

\item{id2geneName}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}

\item{featureType}{either "gene_id" or "transcript_id" or "gene.name" (no collapse function is executed) (default "gene_id")}

\item{...}{Other parameters for scaleTPM}
}
\description{
Classify matrix of RNAseq sample counts by Cell Origin, either original Reddy implementation
or reference-based single sample classification
}
\examples{
\dontrun{
# Classify query samples relative to reference dataset
subType <- computeCOO(query = queryMatrix, useReference = T)
}
}
