% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOriginNet.R
\name{computeOriginNet}
\alias{computeOriginNet}
\title{originNet refined COO (rCOO)}
\usage{
computeOriginNet(query, id2geneName = NULL)
}
\arguments{
\item{query}{Matrix of input samples (values must be in TPM, one row per transcript/gene id)
transcript/gene id). If not specified, will default to Robust reference dataset.}

\item{id2geneName}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}
}
\description{
Novel single-sample COO+DZ classification using internally designed originNet model
}
\examples{
\dontrun{
# Classify query samples with origiNet
subType <- computeOriginNet(query = queryMatrix)
}
}
