% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTME.R
\name{computeTME}
\alias{computeTME}
\title{Classify Sample's TME}
\usage{
computeTME(query, id2geneName = NULL, ...)
}
\arguments{
\item{query}{Matrix of input samples (values must be in TPM, one row per transcript/gene id)
transcript/gene id). If not specified, will default to Robust reference dataset.}

\item{id2geneName}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}

\item{...}{Other parameters for scaleTPM}

\item{featureType}{either "gene_id" or "transcript_id" or "gene.name" (no collapse function is executed) (default "gene_id")}
}
\description{
Classify matrix of RNAseq sample counts by TME26
}
\examples{
\dontrun{
# Classify query samples relative to reference dataset
subType <- computeTME(query = queryMatrix)
}
}
