% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneName.map}
\alias{geneName.map}
\title{Transcript/gene name map}
\format{
A data frame with 199169 rows and 3 variables:
\describe{
\item{transcript_id}{Ensembl transcript ID}
\item{gene_name}{Gene name}
\item{gene_id}{Ensembl gene ID}
}
}
\usage{
geneName.map
}
\description{
Transcript/gene name map
}
\keyword{datasets}
