% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassFromDZSig.R
\name{getClassFromDZSig}
\alias{getClassFromDZSig}
\title{DZSig Class from DZSig score}
\usage{
getClassFromDZSig(score)
}
\arguments{
\item{score}{Numeric vector of DZSig scores}

\item{...}{Other parameters}
}
\description{
DZSig Class from DZSig score
}
