% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassFromNanostring.R
\name{getClassFromNanostring}
\alias{getClassFromNanostring}
\title{COO Class from Nanostring score}
\usage{
getClassFromNanostring(score)
}
\arguments{
\item{score}{Numeric vector of Nanostring scores}
}
\description{
COO Class from Nanostring score
}
