% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelM3.coefficients}
\alias{modelM3.coefficients}
\title{TME26 model coefficients}
\format{
A named vector of 26 coefficients, including the intercept
}
\usage{
modelM3.coefficients
}
\description{
TME26 model coefficients
}
\keyword{datasets}
