% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileNormalizeToRef.R
\name{quantileNormalizeToRef}
\alias{quantileNormalizeToRef}
\title{single-sample quantile normalization}
\usage{
quantileNormalizeToRef(query, ref)
}
\arguments{
\item{query}{Matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{ref}{Numeric vector of reference distribution (one value per gene)
transcript/gene id). If not specified, will default to Robust reference dataset.
\dontrun{
# quantile normalize query counts/TPMs to reference distribution counts/TPMs
query.quant <- quantileNormalizeToRef(query, ref)
}}
}
\description{
single-sample quantile normalization to reference distribution
}
