% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robust.mean.tpm}
\alias{robust.mean.tpm}
\title{ROBUST reference distribution means}
\format{
A named vector of 56478 reference gene means (after tpm scaling)
}
\usage{
robust.mean.tpm
}
\description{
The means of the ROBUST reference RNAseq dataset
}
\keyword{datasets}
