% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robust.ref.mean}
\alias{robust.ref.mean}
\title{ROBUST reference distribution means}
\format{
A named vector of 20 reference gene means
}
\usage{
robust.ref.mean
}
\description{
The means of the ROBUST reference RNAseq dataset
}
\keyword{datasets}
