% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robust.ref.sd}
\alias{robust.ref.sd}
\title{ROBUST reference distribution standard deviations:}
\format{
A named vector of 20 reference gene means
}
\usage{
robust.ref.sd
}
\description{
The standard deviations of the ROBUST reference RNAseq dataset
}
\keyword{datasets}
