% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robust.sd.tpm}
\alias{robust.sd.tpm}
\title{ROBUST reference distribution standard deviations}
\format{
A named vector of 56478 reference gene sds (after tpm scaling)
}
\usage{
robust.sd.tpm
}
\description{
The standard deviations of the ROBUST reference RNAseq dataset
}
\keyword{datasets}
