% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runReddyCOO.R
\name{runReddyCOO}
\alias{runReddyCOO}
\title{Classify sample COO using Reddy algorithm}
\usage{
runReddyCOO(exprs)
}
\arguments{
\item{exprs}{matrix of normalized exprs counts for each sample}
}
\description{
Classify sample COO as GCB-like to ABC-like
}
