% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTPM.R
\name{scaleTPM}
\alias{scaleTPM}
\title{Normalize TPM based on housekeeping genes and self-scale or external ref scale}
\usage{
scaleTPM(
  tpm.mat,
  houseKeeping = c("ISY1", "R3HDM1", "TRIM56", "UBXN4", "WDR55"),
  ref.mean = NULL,
  ref.sd = NULL
)
}
\arguments{
\item{tpm.mat}{matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{houseKeeping}{vector of housekeeping genes to normalize to (ISY1, R3HDM1, TRIM56, UBXN4, WDR55)}

\item{ref.mean, ref.sd}{reference gene means or standard deviations}
}
\description{
Normalize tpm sample counts based on housekeeping genes and self-scale or external ref scale
}
