% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssNormalize.R
\name{ss.normalize}
\alias{ss.normalize}
\title{Normalize sample based on housekeeping genes}
\usage{
ss.normalize(
  tpm.mat,
  id2geneName,
  collapse = TRUE,
  featureType = "gene",
  ref.mean = NULL,
  ref.sd = NULL,
  target = NULL,
  houseKeeping = c("ISY1", "R3HDM1", "TRIM56", "UBXN4", "WDR55"),
  toScale = FALSE
)
}
\arguments{
\item{tpm.mat}{matrix of input samples}

\item{id2geneName}{reference transcripts?}

\item{collapse}{T/F whether to collapse transcripts to gene level (default T)}

\item{featureType}{default "gene"}

\item{ref.mean, ref.sd}{reference gene means or standard deviations}

\item{target}{target}

\item{houseKeeping}{vector of housekeeping genes to normalize to (ISY1, R3HDM1, TRIM56, UBXN4, WDR55)}

\item{toScale}{T/F whether to scale (default F)}
}
\description{
Normalize sample based on housekeeping gene levels
}
