% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssREFERENCE.R
\name{ssREFERENCE}
\alias{ssREFERENCE}
\title{Classify sample COO using Reddy algorithm}
\usage{
ssREFERENCE(query, reference.mean = ref.mean, reference.sd = ref.sd)
}
\arguments{
\item{query}{matrix of normalized query gene/transcript counts for each sample}

\item{reference.mean}{vector of reference means (provided in package data)}

\item{reference.sd}{vector of reference standard deviations (provided in package data)}
}
\description{
Classify sample COO as GCB-like to ABC-like. Note that this method uses a reference
dataset and classifies each sample relative to the reference
}
