% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeClassifier.R
\name{tmeClassifier}
\alias{tmeClassifier}
\title{Apply TME26 classifier}
\usage{
tmeClassifier(exprs)
}
\arguments{
\item{exprs}{Input standardized expression data}
}
\description{
Classify samples according to TME26 classifier
}
