% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseToGenes.R
\name{collapseToGenes}
\alias{collapseToGenes}
\title{Collapse transcript_ids or gene_ids to gene name level}
\usage{
collapseToGenes(tpm.mat, bm = NULL, featureType = "gene_id")
}
\arguments{
\item{tpm.mat}{matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{bm}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}

\item{featureType}{either "gene_id", "transcript_id" (default "gene_id")}
}
\description{
Collapse transcript_ids or gene_ids to gene name level
}
