% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGPclassifiers-package.R
\docType{package}
\name{LGPclassifiers-package}
\alias{LGPclassifiers}
\alias{LGPclassifiers-package}
\title{LGPclassifiers: Classifiers for the LGP Cohort}
\description{
Classifiers for the LGP Cohort. Includes Reddy Cell of Origin single-sample classifier.
}
\author{
\strong{Maintainer}: Clara Amorosi \email{Clara.Amorosi@bms.com} (\href{https://orcid.org/0000-0002-0730-2576}{ORCID})

Authors:
\itemize{
  \item Manuel Ugidos Guerrero \email{Manuel.UgidosGuerrero@bms.com}
  \item Matt Stokes \email{Matt.Stokes@bms.com}
}

Other contributors:
\itemize{
  \item Charalampos Kyriakopoulos \email{Charalampos.Kyriakopoulos@bms.com} [contributor]
}

}
\keyword{internal}
