% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileNormalizeToRef.R
\name{quantileNormalizeToRef}
\alias{quantileNormalizeToRef}
\title{single-sample quantile normalization}
\usage{
quantileNormalizeToRef(target, ref)
}
\arguments{
\item{query}{Matrix of input samples (values must be in TPM, one row per transcript/gene id)
transcript/gene id). If not specified, will default to Robust reference dataset.}

\item{...}{Other parameters for computeOriginNet
\dontrun{
# quantile normalize target counts/TPMs to reference distribution counts/TPMs
normalized_tpm <- quantileNormalizeToRef(target, ref)
}}
}
\description{
single-sample quantile normalization to reference distribution
}
