% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssNormalize.R
\name{ss.normalize}
\alias{ss.normalize}
\title{Normalize sample based on housekeeping genes}
\usage{
ss.normalize(
  tpm.mat,
  id2geneName = NULL,
  collapse = TRUE,
  featureType = "gene",
  houseKeeping = c("ISY1", "R3HDM1", "TRIM56", "UBXN4", "WDR55"),
  ref.mean = NULL,
  ref.sd = NULL
)
}
\arguments{
\item{tpm.mat}{matrix of input samples (values must be in TPM, one row per transcript/gene id)}

\item{id2geneName}{data frame mapping Ensembl gene ids/transcript ids to gene names (provided as internal package data)}

\item{collapse}{T/F whether to collapse transcripts to gene level (default T)}

\item{featureType}{either "gene" or "transcript" (default "gene")}

\item{houseKeeping}{vector of housekeeping genes to normalize to (ISY1, R3HDM1, TRIM56, UBXN4, WDR55)}

\item{ref.mean, ref.sd}{reference gene means or standard deviations}
}
\description{
Normalize tpm sample counts based on housekeeping gene levels
}
