% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssREFERENCE.R
\name{ssREFERENCE}
\alias{ssREFERENCE}
\title{Classify sample COO using Reddy algorithm}
\usage{
ssREFERENCE(query, ref.mean = ref.mean, ref.sd = ref.sd)
}
\arguments{
\item{query}{vector of normalized query gene}

\item{ref.mean}{vector of reference means}

\item{ref.sd}{vector of reference standard deviations}
}
\description{
Classify sample COO. Note this method is implicitly Z-score based and is not "single sample".
It expects a "representative" distribution of COO classes for class thresholds
to work as expected, but scores can be used to rank from GCB-like to ABC-like regardless.
}
